LayoutTests/platform/gtk/http/tests/pasteboard/copy-image-from-context-menu-expected.txt

This tests that using the 'Copy Image' context menu item codepath does not copy the image URL.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.clipboardData.items.length is 2
FAIL event.clipboardData.items[0].type should be text/html. Was text/uri-list.
FAIL event.clipboardData.items[1].type should be image/png. Was text/html.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
To run this test manually, click "Copy Image" in the context menu with the image below, and paste in the box with the black border.

NOTE: This test needs to run on a HTTP server as this bug reproduces only with hosted images.
